Merge in ICMCLCWB/main from autodeploy.v2hito16 to master

* commit 'd36dc1a4e40a8b5a1b27186b85a7733d66d20ae3':
  >Subida pom
  >Subida pom
  >Cambio para recuperar localizacion por ambito
  >Cambio para recuperar localizacion por ambito
  >Coeficiente reduccion jornada
  >Coeficiente reduccion jornada

+++ APPWEB-INCOMEMotorCalculoJava-api/src/main/java/com/inditex/rrhh/icmclcwb/api/app/tarea/dto/TareaPersonaCoeficienteDto.java
@@ -1,6 +1,7 @@
 package com.inditex.rrhh.icmclcwb.api.app.tarea.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 import io.swagger.annotations.ApiModelProperty;
@@ -46,6 +47,6 @@ public class TareaPersonaCoeficienteDto implements Serializable {
 
   @NotNull
   @ApiModelProperty(value = "Coeficiente jornada", required = true)
-  private Long coeficiente;
+  private BigDecimal coeficiente;
 
 }
